+++ CHANGELOG.md
@@ -14,11 +14,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Improved Webp images support with Cloudflare
 - Enable Gzip static to serve precompressed assets with Cache-Enabler and WP-Rocket
 - Stack upgrade for adminer
-
+- Check acme.sh installation and setup acme.sh if needed before issuing certificate
 
 #### Changed
 
 - Add private ip range into fail2ban ignoreip directive
+- Previous `--php73` & `--php73=off` flags are replaced by `--php72`, `--php73`, `--php74` to switch site's php version
 
 #### Fixed